# Make AbstractRowBlock#copyPositions branchless