@lukasz-stec 2022-06-23
> @lukasz-stec, please include the error bars whenever showing benchmark results. It's important to understand the variance to see if the results are significant.

sure, full jmh results for future reference below
```
before
Benchmark                             (nullsAllowed)  (selectedPositions)  (selectedPositionsCount)       (type)  Mode  Cnt   Score   Error  Units
BenchmarkCopyPositions.copyPositions           false              GROUPED                       200  ROW(BIGINT)  avgt   20   0.484 ± 0.003  us/op
BenchmarkCopyPositions.copyPositions           false              GROUPED                      1000  ROW(BIGINT)  avgt   20   2.082 ± 0.015  us/op
BenchmarkCopyPositions.copyPositions           false              GROUPED                      8000  ROW(BIGINT)  avgt   20  18.698 ± 0.116  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                       200  ROW(BIGINT)  avgt   20   0.484 ± 0.004  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                      1000  ROW(BIGINT)  avgt   20   2.097 ± 0.016  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                      8000  ROW(BIGINT)  avgt   20  18.881 ± 0.190  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                       200  ROW(BIGINT)  avgt   20   0.489 ± 0.004  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                      1000  ROW(BIGINT)  avgt   20   2.181 ± 0.091  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                      8000  ROW(BIGINT)  avgt   20  21.120 ± 0.946  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                       200  ROW(BIGINT)  avgt   20   0.862 ± 0.047  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                      1000  ROW(BIGINT)  avgt   20   3.968 ± 0.281  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                      8000  ROW(BIGINT)  avgt   20  41.909 ± 1.993  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                       200  ROW(BIGINT)  avgt   20   0.902 ± 0.010  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                      1000  ROW(BIGINT)  avgt   20   4.389 ± 0.073  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                      8000  ROW(BIGINT)  avgt   20  41.143 ± 2.505  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                       200  ROW(BIGINT)  avgt   20   0.900 ± 0.004  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                      1000  ROW(BIGINT)  avgt   20   4.556 ± 0.144  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                      8000  ROW(BIGINT)  avgt   20  41.507 ± 2.424  us/op

after
Benchmark                             (nullsAllowed)  (selectedPositions)  (selectedPositionsCount)       (type)  Mode  Cnt   Score   Error  Units
BenchmarkCopyPositions.copyPositions           false              GROUPED                       200  ROW(BIGINT)  avgt   20   0.484 ± 0.006  us/op
BenchmarkCopyPositions.copyPositions           false              GROUPED                      1000  ROW(BIGINT)  avgt   20   2.083 ± 0.014  us/op
BenchmarkCopyPositions.copyPositions           false              GROUPED                      8000  ROW(BIGINT)  avgt   20  18.893 ± 0.261  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                       200  ROW(BIGINT)  avgt   20   0.492 ± 0.006  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                      1000  ROW(BIGINT)  avgt   20   2.102 ± 0.008  us/op
BenchmarkCopyPositions.copyPositions           false             SEQUENCE                      8000  ROW(BIGINT)  avgt   20  18.914 ± 0.173  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                       200  ROW(BIGINT)  avgt   20   0.492 ± 0.003  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                      1000  ROW(BIGINT)  avgt   20   2.145 ± 0.022  us/op
BenchmarkCopyPositions.copyPositions           false               RANDOM                      8000  ROW(BIGINT)  avgt   20  20.017 ± 0.904  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                       200  ROW(BIGINT)  avgt   20   0.894 ± 0.048  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                      1000  ROW(BIGINT)  avgt   20   4.005 ± 0.214  us/op
BenchmarkCopyPositions.copyPositions            true              GROUPED                      8000  ROW(BIGINT)  avgt   20  27.061 ± 1.552  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                       200  ROW(BIGINT)  avgt   20   0.940 ± 0.021  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                      1000  ROW(BIGINT)  avgt   20   4.174 ± 0.111  us/op
BenchmarkCopyPositions.copyPositions            true             SEQUENCE                      8000  ROW(BIGINT)  avgt   20  28.406 ± 0.198  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                       200  ROW(BIGINT)  avgt   20   0.927 ± 0.007  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                      1000  ROW(BIGINT)  avgt   20   4.170 ± 0.131  us/op
BenchmarkCopyPositions.copyPositions            true               RANDOM                      8000  ROW(BIGINT)  avgt   20  29.262 ± 0.159  us/op
```